data_quality_monitor/pyproject.toml
[tool.poetry]
name = "data-quality-monitor"
version = "0.1.0"
description = "Data quality monitoring service"
authors = ["Andrew Ghostuhin <me@torinasakura.name>"]
packages = [{include = "data_quality_monitor", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
fastapi = "^0.120.4"
uvicorn = {version = "^0.38", extras = ["standard"]}
pandas = "^2.2"
numpy = "^2.3.4"
pyyaml = "^6.0"
clickhouse-connect = "^0.9.2"
loguru = "^0.7"
prometheus-client = "^0.23.1"
typer = {version = "^0.20", extras = ["all"]}
confluent-kafka = "^2.12.0"
starlette = ">=0.40.0,<0.50.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2"
vulture = "^2.14"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 150
select = ["E", "F", "W", "C90"]
